requirements.txt
openai
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
prompt_toolkit
